Add material-web and lit
This commit adds a duration dropdown to the channel view.

I chose to go with material design for the select component and
discovered that the library uses lit, which seems cool.

Had to do some lit and material theme magic to get things to work,
but it seems to be working.