Use channel duration for on function
This commit sets lastOnDuration from the current channel duration
when setting the channel to on.

The on and off buttons have been combined into the main it's on button.

Loading state has been synced to all buttons on the channel component.

All em styles have been updated to rem for consistency's sake.